Refactored JobDTO

+++ Microservice/JobMicroService/src/main/java/com/numan947/JobMicroService/job/impl/JobServiceImpl.java
@@ -5,6 +5,8 @@ import com.numan947.JobMicroService.job.JobModel;
 import com.numan947.JobMicroService.job.JobRepository;
 import com.numan947.JobMicroService.job.JobService;
 import com.numan947.JobMicroService.job.dto.JobWithCompanyDTO;
+import com.numan947.JobMicroService.job.mapper.JobMapper;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
@@ -15,9 +17,11 @@ import java.util.stream.Collectors;
 @Service
 public class JobServiceImpl implements JobService {
     JobRepository jobRepository;
+    final RestTemplate loadBalancedRestTemplate;
 
-    public JobServiceImpl(JobRepository jobRepository) {
+    public JobServiceImpl(JobRepository jobRepository, RestTemplate loadBalancedRestTemplate) {
         this.jobRepository = jobRepository;
+        this.loadBalancedRestTemplate = loadBalancedRestTemplate;
     }
 
     @Override
@@ -26,8 +30,8 @@ public class JobServiceImpl implements JobService {
     }
 
     @Override
-    public JobModel findJobById(Long id) {
-        return jobRepository.findById(id).orElse(null);
+    public JobWithCompanyDTO findJobById(Long id) {
+        return getJobWithCompanyDTO(jobRepository.findById(id).orElse(null));
     }
 
     @Override
@@ -39,13 +43,18 @@ public class JobServiceImpl implements JobService {
     }
 
     @Override
-    public void deleteJob(Long id) {
+    public boolean deleteJob(Long id) {
+        JobModel j = jobRepository.findById(id).orElse(null);
+        if (j == null){
+            return false;
+        }
         jobRepository.deleteById(id);
+        return true;
     }
 
     @Override
-    public void updateJob(Long id, JobModel job) {
-        JobModel j = findJobById(id);
+    public boolean updateJob(Long id, JobModel job) {
+        JobModel j = jobRepository.findById(id).orElse(null);
         if (j != null) {
             j.setTitle(job.getTitle());
             j.setDescription(job.getDescription());
@@ -53,15 +62,16 @@ public class JobServiceImpl implements JobService {
             j.setMaxSalary(job.getMaxSalary());
             j.setLocation(job.getLocation());
             jobRepository.save(j);
+            return true;
         }
+        return false;
     }
 
     private JobWithCompanyDTO getJobWithCompanyDTO(JobModel j){
         if (j == null){
             return null;
         }
-        RestTemplate rt = new RestTemplate();
-        Company c = rt.getForObject("http://localhost:8081/companies/" + j.getCompanyId(), Company.class);
-        return new JobWithCompanyDTO(j, c);
+        Company c = loadBalancedRestTemplate.getForObject("http://CompanyMicroService:8081/companies/" + j.getCompanyId(), Company.class);
+        return JobMapper.toJobWithCompanyDTO(j, c);
     }
 }
